Move pkg/jtracer to internal/jtracer
