+++ internal/storage/v1/cassandra/savetracetest/main.go
@@ -15,7 +15,7 @@ import (
 	"github.com/jaegertracing/jaeger/internal/storage/v1/cassandra"
 	cSpanStore "github.com/jaegertracing/jaeger/internal/storage/v1/cassandra/spanstore"
 	cascfg "github.com/jaegertracing/jaeger/pkg/cassandra/config"
-	"github.com/jaegertracing/jaeger/pkg/jtracer"
+	"github.com/jaegertracing/jaeger/internal/jtracer"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 )
